+++ src/Atata.WebDriverSetup/Strategies/FirefoxDriverSetupStrategy.cs
@@ -52,7 +52,7 @@ public class FirefoxDriverSetupStrategy :
                 ?? RegistryUtils.GetValue(@"HKEY_CURRENT_USER\Software\Mozilla\Mozilla Firefox")
                 ?? AppVersionDetector.GetByApplicationPathInRegistry("firefox.exe")
             : (OSInfo.IsMacOS
-                ? (await AppVersionDetector.GetThroughOSXApplicationCliAsync("Firefox", cancellationToken).ConfigureAwait(false))
+                ? (await AppVersionDetector.GetThroughMacOSApplicationCliAsync("Firefox", cancellationToken).ConfigureAwait(false))
                 : (await AppVersionDetector.GetThroughCliAsync("firefox", "-v", cancellationToken).ConfigureAwait(false)))
                 ?.Replace("Mozilla Firefox ", null);